HSEARCH-3881 Rename a few tests whose name included a typo
